+++ lib/src/widgets/raw_editor.dart
@@ -363,7 +363,7 @@ class RawEditorState extends EditorState
           pasteEnabled ? () => pasteText(SelectionChangedCause.toolbar) : null,
       onSelectAll: selectAllEnabled
           ? () => selectAll(SelectionChangedCause.toolbar)
-          : null,
+          : null, onLiveTextInput: () {  },
     );
   }
 
@@ -1752,6 +1752,10 @@ class RawEditorState extends EditorState
       }
     }
   }
+
+  @override
+  // TODO: implement liveTextInputEnabled
+  bool get liveTextInputEnabled => throw UnimplementedError();
 }
 
 class _Editor extends MultiChildRenderObjectWidget {