+++ CHANGELOG.md
@@ -1,5 +1,18 @@
+# [7.2.18]
+- Fix paste block text in words apply same style.
+
+# [7.2.17]
+- Fix paste text mess up style.
+- Add support copy/cut block text.
+
+# [7.2.16]
+- Allow for custom context menu.
+
+# [7.2.15]
+- Add flutter_quill.delta library which only exposes Delta datatype.
+
 # [7.2.14]
-- Fix errors when the editor is used in the `screenshot` package
+- Fix errors when the editor is used in the `screenshot` package.
 
 # [7.2.13]
 - Fix around image can't delete line break.